+++ Tests/Plugin/FakeIpPluginTest.php
@@ -33,6 +33,21 @@ class FakeIpPluginTest extends TestCase
         $this->assertSame($query->getText(), '123.123.123.123');
     }
 
+    /**
+     * @testWith [null]
+     *           [""]
+     */
+    public function testEmptyLocalIpQuery(?string $localIp)
+    {
+        $fakeIpPlugin = new FakeIpPlugin($localIp, '123.123.123.123');
+        $query = GeocodeQuery::create('124.124.124.124');
+
+        /** @var Query $query */
+        $query = $fakeIpPlugin->handleQuery($query, function (Query $query) { return $query; }, function () {});
+
+        $this->assertSame($query->getText(), '123.123.123.123');
+    }
+
     public function testHandleQueryUsingFaker()
     {
         $fakeIpPlugin = new FakeIpPlugin('127.0.0.1', '192.168.1.1', true);