+++ Tests/bootstrap.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  */
 
 if (!$loader = @require_once __DIR__.'/../vendor/autoload.php') {
-    die('You must set up the project dependencies, run the following commands:
+    exit('You must set up the project dependencies, run the following commands:
 wget http://getcomposer.org/composer.phar
 php composer.phar install --dev --prefer-source
 ');